Update text-cairo.c
The modification at lines 794 thru 824 fixes the issue of centering the text for the 2nd line of a button that has 2 lines of text.  The modification at lines 595 thru 607 fixes the issue of breaking and centering a long word to fit on a 2 line button.  This happens commonly with Japanese character strings, where the string contains no spaces.